refactor: S3ImageUpload를 Component로 변경
