+++ src/main/java/com/ctrls/auto_enter_view/service/ResumeImageService.java
@@ -1,5 +1,6 @@
 package com.ctrls.auto_enter_view.service;
 
+import com.ctrls.auto_enter_view.component.S3ImageUpload;
 import com.ctrls.auto_enter_view.dto.resume.ResumeDto;
 import com.ctrls.auto_enter_view.entity.ResumeImageEntity;
 import com.ctrls.auto_enter_view.repository.ResumeImageRepository;
@@ -18,7 +19,7 @@ import org.springframework.web.multipart.MultipartFile;
 public class ResumeImageService {
 
   private final ResumeImageRepository resumeImageRepository;
-  private final S3ImageUploadService s3ImageUploadService;
+  private final S3ImageUpload s3ImageUpload;
   private final ResumeRepository resumeRepository;
 
   /**
@@ -37,10 +38,10 @@ public class ResumeImageService {
 
     // 기존 이미지가 있고, 새 이미지가 제공된 경우에만 기존 이미지 삭제
     if (resumeImage.getResumeImageUrl() != null) {
-      s3ImageUploadService.deleteImage(resumeImage.getResumeImageUrl());
+      s3ImageUpload.deleteImage(resumeImage.getResumeImageUrl());
     }
 
-    String imageUrl = s3ImageUploadService.uploadImage(image, "resume-images");
+    String imageUrl = s3ImageUpload.uploadImage(image, "resume-images");
     resumeImage.updateResumeImageUrl(imageUrl);
     resumeImageRepository.save(resumeImage);
 
@@ -75,7 +76,7 @@ public class ResumeImageService {
           resumeImageRepository.findByResumeKey(resumeKey)
               .ifPresent(image -> {
                 String imageUrl = image.getResumeImageUrl();
-                s3ImageUploadService.deleteImage(imageUrl);
+                s3ImageUpload.deleteImage(imageUrl);
                 resumeImageRepository.delete(image);
               });
         });
